Refactor component imports and paths
