+++ frontend/src/components/call/PassButton.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import styled from "styled-components";
-import { theme } from "./theme";
+import { theme } from "../../utils/theme";
 
 type Props = {
   $calling: boolean;
