+++ frontend/src/components/qr/QrCode.tsx
@@ -1,7 +1,7 @@
 import { useQRCode } from "next-qrcode";
 import React, { FC } from "react";
 import styled from "styled-components";
-import { theme } from "./theme";
+import { theme } from "../../utils/theme";
 
 type Props = {
   url: string;
