+++ frontend/src/pages/startPage.tsx
@@ -1,7 +1,7 @@
 import { FC } from "react";
 import styled from "styled-components";
-import { Button } from "../utils/Button";
-import { MessageCricle } from "../utils/MessageCricle";
+import { Button } from "../components/utils/Button";
+import { MessageCricle } from "../components/utils/MessageCricle";
 
 const StartPageContainer = styled.div`
   text-align: center;
